+++ galsim/chromatic.py
@@ -308,12 +308,7 @@ class ChromaticObject(object):
         # setup output image using fiducial profile
         wave0, prof0 = self._fiducial_profile(bandpass)
         image = prof0.drawImage(image=image, setup_only=True, **kwargs)
-        # Remove from kwargs anything that is only used for setting up image:
-        kwargs.pop('dtype', None)
-        kwargs.pop('scale', None)
-        kwargs.pop('wcs', None)
-        kwargs.pop('nx', None)
-        kwargs.pop('ny', None)
+        _remove_setup_kwargs(kwargs)
 
         # determine combined self.wave_list and bandpass.wave_list
         wave_list = self._getCombinedWaveList(bandpass)
@@ -911,12 +906,7 @@ class InterpolatedChromaticObject(ChromaticObject):
         # and so on.
         _, prof0 = self._fiducial_profile(bandpass)
         image = prof0.drawImage(image=image, setup_only=True, **kwargs)
-        # Remove from kwargs anything that is only used for setting up image:
-        kwargs.pop('dtype', None)
-        kwargs.pop('scale', None)
-        kwargs.pop('wcs', None)
-        kwargs.pop('nx', None)
-        kwargs.pop('ny', None)
+        _remove_setup_kwargs(kwargs)
 
         # determine combination of self.wave_list and bandpass.wave_list
         wave_list = self._getCombinedWaveList(bandpass)
@@ -1597,6 +1587,7 @@ class ChromaticSum(ChromaticObject):
         # Use given add_to_image for the first one, then add_to_image=False for the rest.
         image = self.objlist[0].drawImage(
                 bandpass, image=image, add_to_image=add_to_image, **kwargs)
+        _remove_setup_kwargs(kwargs)
         for obj in self.objlist[1:]:
             image = obj.drawImage(
                     bandpass, image=image, add_to_image=True, **kwargs)
@@ -1874,6 +1865,7 @@ class ChromaticConvolution(ChromaticObject):
                     tmplist.append(summand)
                     tmpobj = ChromaticConvolution(tmplist)
                     # add to previously started image
+                    _remove_setup_kwargs(kwargs)
                     image = tmpobj.drawImage(bandpass, image=image, integrator=integrator,
                                              iimult=iimult, add_to_image=True, **kwargs)
                 # Return the image here, breaking the loop early.  If there are two ChromaticSum
@@ -1887,12 +1879,7 @@ class ChromaticConvolution(ChromaticObject):
         # setup output image (semi-arbitrarily using the bandpass effective wavelength)
         wave0, prof0 = self._fiducial_profile(bandpass)
         image = prof0.drawImage(image=image, setup_only=True, **kwargs)
-        # Remove from kwargs anything that is only used for setting up image:
-        kwargs.pop('dtype', None)
-        kwargs.pop('scale', None)
-        kwargs.pop('wcs', None)
-        kwargs.pop('nx', None)
-        kwargs.pop('ny', None)
+        _remove_setup_kwargs(kwargs)
 
         # Sort these atomic objects into separable and inseparable lists, and collect
         # the spectral parts of the separable profiles.
@@ -2184,8 +2171,8 @@ class ChromaticAiry(ChromaticObject):
         if diam is not None:
             self.lam_over_diam = (1.e-9*lam/diam)*galsim.radians/scale_unit
         else:
-            self.lam_over_diam = lam_over_diam
-        self.lam = lam
+            self.lam_over_diam = float(lam_over_diam)
+        self.lam = float(lam)
 
         self.kwargs = kwargs
         self.scale_unit = scale_unit
@@ -2241,3 +2228,15 @@ def _linearInterp(list, frac, lower_idx):
     interpolation later on if we want to enable something other than linear interpolation.
     """
     return frac*list[lower_idx+1] + (1.-frac)*list[lower_idx]
+
+def _remove_setup_kwargs(kwargs):
+    """
+    Helper function to remove from kwargs anything that is only used for setting up image and that
+    might otherwise interfere with drawImage.
+    """
+    kwargs.pop('dtype', None)
+    kwargs.pop('scale', None)
+    kwargs.pop('wcs', None)
+    kwargs.pop('nx', None)
+    kwargs.pop('ny', None)
+    kwargs.pop('bounds', None)